backend/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.5
django-environ==0.11.2
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
hiredis==2.2.3
idna==3.4
Markdown==3.4.4
psycopg2-binary==2.9.7
PyJWT==2.8.0
pytz==2023.3.post1
redis==5.0.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
urllib3==2.0.4
